completed seed and model

+++ data/migrations/01-make_cars_table.js
@@ -4,7 +4,7 @@ exports.up = function (knex) {
     tbl.string('VIN', 17).unique().notNullable()//required, unique
     tbl.string('Make', 150).notNullable()//required
     tbl.string('Model', 150).notNullable()//required
-    tbl.numeric('Mileage').notNullable()//required
+    tbl.float('Mileage').notNullable()//required
     tbl.string('Title', 150) //optional
     tbl.string('Transmission', 150) //optional
   })
